fix(phantom): ignore shutdown on cleanly exited procs

+++ src/phantom.js
@@ -34,6 +34,10 @@ export default class Phantom {
   }
 
   shutdown() {
+    if (this.phantom === null || this.phantom === undefined) {
+      return;
+    }
+
     this.phantom.stdin.end();
     this.phantom.kill();
   }
